pyproject.toml
[tool.poetry]
name = "osod-impl"
version = "0.1.0"
description = "Open-Set Object Detection scaffold (Faster R-CNN + CLIP text encoder)"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
# Constrain to 3.11.x (exclude 3.12+) for wheel availability.
python = ">=3.11,<3.12"
torch = "2.4.1"
torchvision = "0.19.1"
numpy = "^2.4.2"
pycocotools = "^2.0.8"
open-clip-torch = "^2.24.0"
tqdm = "^4.67.2"
datasets = "^4.5.0"
kaggle = "^1.8.4"

[tool.poetry.group.dev.dependencies]
black = ">=24.4.0,<25.0.0"
ruff = ">=0.6.0,<0.7.0"
mypy = ">=1.8.0,<1.12.0"
pytest = ">=7.4,<9.0"
pytest-cov = ">=4.1,<6.0"
pre-commit = ">=3.6,<4.0"
python-dotenv = ">=1.0,<2.0"

[tool.black]
line-length = 88
target-version = ["py311"]
extend-exclude = """
(?x)(
    ^data/
  | ^.venv/
  | __pycache__/
  | \\.ipynb$
)
"""

[tool.ruff]
target-version = "py311"
line-length = 88
extend-exclude = [
  "data",
  ".venv",
  "__pycache__",
  "*.ipynb",
]

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "F",   # pyflakes
  "W",   # pycodestyle warnings
  "B",   # flake8-bugbear
  "C4",  # flake8-comprehensions
  "I",   # isort
  "UP",  # pyupgrade
  "SIM", # flake8-simplify
]
ignore = [
  "E203",
  "E266",
]

[tool.ruff.lint.isort]
combine-as-imports = true
force-sort-within-sections = true

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_unused_ignores = true
warn_unused_configs = true
show_error_codes = true
pretty = true
exclude = """
(?x)(
    ^data/
  | ^.venv/
  | __pycache__/
  | \\.ipynb$
)
"""

